requirements.txt
celery
django
djangorestframework
django-debug-toolbar
django-extensions
django-resized
Pillow
psycopg2
